refactor: move DiskSpaceInfo to system_monitor to reduce coupling
Moved DiskSpaceInfo from panels/file_manager to system_monitor module, where it logically belongs alongside CPU and RAM monitoring. This eliminates cross-panel dependency where terminal_pty imported from file_manager, reducing coupling and improving module cohesion.

Benefits:
- Terminal no longer depends on file_manager module
- System resource info (CPU, RAM, Disk) now centralized
- Better adherence to Single Responsibility Principle
- Cleaner module boundaries
